+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Added a preference to include (or not) videos in results. Video files (.mp4, .mov, etc) don't display a thumbnail when referenced in a Raycast Grid Item yet, but work as expected when opening the file or using quicklook.
 - Updated Grid Item to display file icon for videos instead of a dot icon.
+- Move Grid itemSize into the extension preferences
+- Use the Grid Dropdown to swap between "Everything" and each of the input folders. By default, the extension will open to show all folders at once.
 
 #### 2022-07-07